+++ CHANGELOG.md
@@ -8,3 +8,6 @@
 01. :gift: Added interface `dev.orne.http.client.StatusDependentOperation`
 01. :gift: Added interface `dev.orne.http.client.StatusInitOperation`
 01. :gift: Added class `dev.orne.http.client.StatedHttpServiceClient`
+01. :gift: Added interface `dev.orne.http.client.AuthenticableClientStatus`
+01. :gift: Added interface `dev.orne.http.client.AuthenticatedOperation`
+01. :gift: Added exception `dev.orne.http.client.HttpAuthenticationRequiredException`